requirements.txt
requests
Django<1.11
django-extensions
django-filter
django-flat-theme
djangorestframework==3.5.3
ipython
psycopg2==2.6
pytest-django==3.1.2
pytest==3.0.5
python-dateutil
pytz
beautifulsoup4
responses
gunicorn
uwsgi
icalendar
opbeat
django-cors-headers==1.1.0
